OneDrive/Desktop/LM Software Solution/google sheet project/project/requirements.txt
fastapi
uvicorn
gspread
google-auth
google-auth-oauthlib
python-dotenv
